move SearchResult components to separate files
