+++ src/components/LinkChooser/SearchResultsList.tsx
@@ -1,21 +1,16 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { getKeyItemRecord, getMenuItems } from "@components/ActionMenu/Aria/helper";
-import { MenuItemContentSize } from "@components/MenuItem";
-import { MenuItem } from "@components/MenuItem/MenuItem";
-import { getInteractionModality } from "@react-aria/interactions";
-import { useListBox, useListBoxSection, useOption } from "@react-aria/listbox";
-import { merge } from "@utilities/merge";
+import { useListBox } from "@react-aria/listbox";
 import { useActor } from "@xstate/react";
 import React, { FC, useMemo, useRef } from "react";
 import BackgroundIcon from "./assets/background.svg";
 import NoResultsIcon from "./assets/no-results.svg";
 import FetchingIcon from "./assets/nook-animated.png";
-import { IconOptions } from "./LinkChooser";
 import { NavigationMenuItem } from "./NavigationMenu";
+import { SearchResultSection } from "./SearchResultSection";
 import { defaultSection } from "./sections";
-import { DropdownState, LinkChooserState, SectionState } from "./state/machine";
-import { ImageMenuItemProps, SearchResultListProps, SearchResultOptionProps, SearchResultSectionProps } from "./types";
+import { SearchResultListProps } from "./types";
 import { findSection } from "./utils/helpers";
 import { isFetching, isUnsuccessful, shouldGoBack } from "./utils/state";
 
@@ -26,11 +21,17 @@ export const SearchResultsList: FC<SearchResultListProps> = (props) => {
     const items = getMenuItems(menuBlocks);
     const keyItemRecord = getKeyItemRecord(items);
 
-    const [{ context, matches }, send] = useActor(machineService);
+    const [
+        {
+            context: { extraSections, currentSectionId, searchResults, query },
+            matches,
+        },
+        send,
+    ] = useActor(machineService);
 
     const currentSection = useMemo(
-        () => findSection(context.extraSections, context.currentSectionId) || defaultSection,
-        [context.currentSectionId],
+        () => findSection(extraSections, currentSectionId) || defaultSection,
+        [currentSectionId],
     );
 
     if (isFetching(matches)) {
@@ -64,7 +65,7 @@ export const SearchResultsList: FC<SearchResultListProps> = (props) => {
                 ref={listBoxRef}
                 className="tw-list-none tw-p-0 tw-m-0 tw-bg-white tw-z-20 focus-visible:tw-outline-none"
             >
-                {context.searchResults.length ? (
+                {searchResults.length ? (
                     [...state.collection]
                         .filter((section) => section.key === "search")
                         .map((item) => (
@@ -79,83 +80,15 @@ export const SearchResultsList: FC<SearchResultListProps> = (props) => {
                 ) : (
                     <EmptyResults
                         prompt={
-                            context.query
-                                ? `We could not find any results for "${context.query}".`
+                            query
+                                ? `We could not find any results for "${query}".`
                                 : "Use the search above to discover your brand assets"
                         }
-                        icon={context.query ? NoResultsIcon : BackgroundIcon}
+                        icon={query ? NoResultsIcon : BackgroundIcon}
                     />
                 )}
             </ul>
         </div>
-    );
-};
-
-const SearchResultSection: FC<SearchResultSectionProps> = ({ heading, state, keyItemRecord, machineService }) => {
-    const { itemProps, groupProps } = useListBoxSection({
-        heading: heading.rendered,
-        "aria-label": heading["aria-label"],
-    });
-
-    return (
-        <li {...itemProps} className="tw-border-b tw-border-b-black-10 last:tw-border-0">
-            <ul
-                {...groupProps}
-                data-test-id="link-chooser-select-section"
-                className="tw-py-2 tw-px-0 tw-m-0 tw-list-none"
-            >
-                {[...heading.childNodes].map((node) => (
-                    <SearchResultOption
-                        key={node.key}
-                        item={node}
-                        state={state}
-                        keyItemRecord={keyItemRecord}
-                        machineService={machineService}
-                    />
-                ))}
-            </ul>
-        </li>
-    );
-};
-
-const SearchResultOption: FC<SearchResultOptionProps> = ({ item, state, keyItemRecord, machineService }) => {
-    const ref = useRef<HTMLLIElement>(null);
-    const { optionProps, isDisabled, isSelected, isFocused } = useOption(
-        {
-            key: item.key,
-            shouldFocusOnHover: false,
-        },
-        state,
-        ref,
-    );
-    const [{ context, matches }] = useActor(machineService);
-
-    const menuItem = keyItemRecord[item.key];
-    const decorator = menuItem.icon ? IconOptions[menuItem.icon] : undefined;
-    const currentSection = findSection(context.extraSections, context.currentSectionId);
-    const isLoaded = (currentState: DropdownState) =>
-        matches(`${LinkChooserState.Focused}.${currentState}.${SectionState.Loaded}`);
-
-    const isFocusVisible = getInteractionModality() !== "pointer";
-
-    return (
-        <li
-            {...optionProps}
-            ref={ref}
-            className={merge([
-                "tw-relative hover:tw-bg-black-0 tw-list-none tw-outline-none",
-                isDisabled && "tw-pointer-events-none tw-top-px",
-                isFocused && isFocusVisible && "tw-bg-black-0",
-            ])}
-        >
-            {isLoaded(DropdownState.Default) ? (
-                <MenuItem {...menuItem} active={isSelected} decorator={decorator} size={MenuItemContentSize.Large} />
-            ) : isLoaded(DropdownState.ExtraSection) && currentSection?.renderPreview ? (
-                currentSection.renderPreview(menuItem)
-            ) : (
-                <ImageMenuItem {...menuItem} />
-            )}
-        </li>
     );
 };
 
@@ -185,17 +118,5 @@ const FetchingAnimation: FC = () => (
         className="tw-flex tw-flex-col tw-justify-center tw-items-center tw-h-[220px] tw-p-3"
     >
         <img className="tw-w-[50px]" src={FetchingIcon} alt="Fetching" />
-    </div>
-);
-
-const ImageMenuItem: FC<ImageMenuItemProps> = ({ title, subtitle, preview }) => (
-    <div className="tw-flex tw-px-5 tw-py-1.5 tw-cursor-pointer">
-        <div className="tw-flex tw-shrink-0 tw-w-[75px] tw-h-[75px] tw-max-w-xs tw-bg-black-0 tw-rounded">
-            {preview && <img className="tw-w-full tw-object-contain" src={preview} alt={title as string} />}
-        </div>
-        <div className="tw-flex tw-flex-col tw-justify-center tw-ml-4">
-            <p className="tw-block tw-text-md tw-leading-tight tw-hover:underline tw-text-black-80">{title}</p>
-            {subtitle && <p className="tw-text-[12px] tw-text-black-80">{subtitle}</p>}
-        </div>
     </div>
 );
